+++ R/dataframe__frame.R
@@ -852,9 +852,9 @@ DataFrame_group_by = function(..., maintain_order = pl$options$maintain_order) {
 #' @examples
 #' df = pl$DataFrame(iris[1:3, ])
 #' df$to_data_frame()
-DataFrame_to_data_frame = function(..., bigint_conversion = pl$options$bigint_conversion) {
+DataFrame_to_data_frame = function(..., int64_conversion  = pl$options$int64_conversion ) {
   # do not unnest structs and mark with I to also preserve categoricals as is
-  l = lapply(self$to_list(unnest_structs = FALSE, bigint_conversion), I)
+  l = lapply(self$to_list(unnest_structs = FALSE, int64_conversion ), I)
 
   # similar to as.data.frame, but avoid checks, whcih would edit structs
   df = data.frame(seq_along(l[[1L]]), ...)
@@ -883,12 +883,12 @@ DataFrame_to_data_frame = function(..., bigint_conversion = pl$options$bigint_co
 #' @keywords DataFrame
 #' @examples
 #' pl$DataFrame(iris)$to_list()
-DataFrame_to_list = function(unnest_structs = TRUE, bigint_conversion = pl$options$bigint_conversion) {
+DataFrame_to_list = function(unnest_structs = TRUE, int64_conversion  = pl$options$int64_conversion ) {
   if (unnest_structs) {
-    .pr$DataFrame$to_list(self, bigint_conversion) |>
+    .pr$DataFrame$to_list(self, int64_conversion ) |>
       unwrap("in $to_list():")
   } else {
-    .pr$DataFrame$to_list_tag_structs(self, bigint_conversion) |>
+    .pr$DataFrame$to_list_tag_structs(self, int64_conversion ) |>
       unwrap("in $to_list():") |>
       restruct_list()
   }