+++ NEWS.md
@@ -14,10 +14,10 @@
 -   New methods `$list$any()` and `$list$all()` (#709).
 -   New function `pl$from_epoch()` to convert a Unix timestamp to a date(time)
     variable (#708).
--   New option `bigint_conversion` to specify how Int64 columns (that don't have
+-   New option `int64_conversion ` to specify how Int64 columns (that don't have
     equivalent in base R) should be converted. This option can either be set 
     globally with `pl$set_options()` or on a case-by-case basis, e.g with 
-    `$to_data_frame(bigint_conversion =)` (#706).
+    `$to_data_frame(int64_conversion  =)` (#706).
 
 ## polars 0.12.2